refactor: move the logic for saving a new user from domain layer to application layer
