--- src/main/java/br/com/aegro/datasync/user/domain/SaveUserService.java
@@ -1,5 +1,0 @@
-package br.com.aegro.datasync.user.domain;
-
-public interface SaveUserService {
-    User execute(User user);
-}
